requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
alembic>=1.13
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic>=2
email-validator>=2
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
python-jose[cryptography]==3.3.0
python-multipart>=0.0.9
Pillow>=10.0.0